--- packages/moonstone/Scroller/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Scroller.js"
-}
