+++ packages/moonstone/VirtualFlexList/VirtualFlexList.js
@@ -12,7 +12,7 @@ import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 
-import {VirtualListCore} from '../VirtualList/VirtualListBase';
+import {VirtualListCore} from '@enact/ui/VirtualList/VirtualListBase';
 
 import css from './VirtualFlexList.less';
 import Positionable from './Positionable';
