+++ packages/sampler/stories/moonstone-stories/VirtualList.js
@@ -1,6 +1,7 @@
 import Item from '@enact/moonstone/Item';
 import VirtualList from '@enact/moonstone/VirtualList';
-import {VirtualListCore} from '@enact/moonstone/VirtualList/VirtualListBase';
+// import {VirtualListNative as VirtualList} from '@enact/moonstone/VirtualList/VirtualListNative.js';
+import {VirtualListCore} from '@enact/ui/VirtualList/VirtualListBase';
 import ri from '@enact/ui/resolution';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
