+++ packages/sampler/stories/qa-stories/VirtualGridList.js
@@ -1,6 +1,6 @@
 import {VirtualGridList} from '@enact/moonstone/VirtualList';
-import GridListImageItem from '@enact/moonstone/VirtualList/GridListImageItem';
-import {VirtualListCore} from '@enact/moonstone/VirtualList/VirtualListBase';
+import GridListImageItem from '@enact/ui/VirtualList/GridListImageItem';
+import {VirtualListCore} from '@enact/ui/VirtualList/VirtualListBase';
 import ri from '@enact/ui/resolution';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
