+++ packages/ui/Scrollable/ScrollableNative.js
@@ -1,30 +1,29 @@
 /*
- * Exports the {@link moonstone/Scroller.Scrollable} Higher-order Component (HOC) and
+ * Exports the {@link moonstone/Scroller.ScrollableNative} Higher-order Component (HOC) and
  * the {@link moonstone/Scroller.dataIndexAttribute} constant.
- * The default export is {@link moonstone/Scroller.Scrollable}.
+ * The default export is {@link moonstone/Scroller.ScrollableNative}.
  */
 
 import clamp from 'ramda/src/clamp';
 import classNames from 'classnames';
 import {contextTypes as contextTypesResize} from '@enact/ui/Resizable';
+import {contextTypes as contextTypesRtl} from '@enact/i18n/I18nDecorator';
 import {forward} from '@enact/core/handle';
 import {getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import hoc from '@enact/core/hoc';
-import {on, off} from '@enact/core/dispatcher';
 import {is} from '@enact/core/keymap';
-import {perfNow} from '@enact/core/util';
+import {Job} from '@enact/core/util';
+import {on, off} from '@enact/core/dispatcher';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import ri from '@enact/ui/resolution';
 import Spotlight from '@enact/spotlight';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
-import {contextTypes as contextTypesState, Publisher} from '@enact/core/internal/PubSub';
 
-import ScrollAnimator from './ScrollAnimator';
-import Scrollbar from './Scrollbar';
+import Scrollbar from '../Scrollbar';
 
-import css from './Scrollable.less';
-import scrollbarCss from './Scrollbar.less';
+import css from '@enact/moonstone/Scroller/Scrollable.less';
+import scrollbarCss from '@enact/moonstone/Scroller/Scrollbar.less';
 
 const
 	forwardScroll = forward('onScroll'),
@@ -32,15 +31,14 @@ const
 	forwardScrollStop = forward('onScrollStop');
 
 const
-	calcVelocity = (d, dt) => (d && dt) ? d / dt : 0,
 	nop = () => {},
-	holdTime = 50,
+	paginationPageMultiplier = 0.8,
+	epsilon = 1,
 	scrollWheelMultiplierForDeltaPixel = 1.5, // The ratio of wheel 'delta' units to pixels scrolled.
 	scrollWheelPageMultiplierForMaxPixel = 0.2, // The ratio of the maximum distance scrolled by wheel to the size of the viewport.
 	pixelPerLine = 39,
-	paginationPageMultiplier = 0.8,
-	epsilon = 1,
-	animationDuration = 1000,
+	scrollStopWaiting = 200,
+	// spotlight
 	isPageUp = is('pageUp'),
 	isPageDown = is('pageDown'),
 	reverseDirections = {
@@ -50,19 +48,7 @@ const
 		'down': 'up'
 	};
 
-/**
- * {@link moonstone/Scroller.dataIndexAttribute} is the name of a custom attribute
- * which indicates the index of an item in {@link moonstone/VirtualList.VirtualList}
- * or {@link moonstone/VirtualList.VirtualGridList}.
- *
- * @constant dataIndexAttribute
- * @memberof moonstone/Scroller
- * @type {String}
- * @private
- */
-const dataIndexAttribute = 'data-index';
-
-const ScrollableSpotlightContainer = SpotlightContainerDecorator(
+const ScrollableSpotlightContainerDecorator = SpotlightContainerDecorator(
 	{
 		navigableFilter: (elem, {focusableScrollbar}) => {
 			if (
@@ -87,25 +73,36 @@ const ScrollableSpotlightContainer = SpotlightContainerDecorator(
 );
 
 /**
- * {@link moonstone/Scroller.Scrollable} is a Higher-order Component
+ * {@link moonstone/Scroller.dataIndexAttribute} is the name of a custom attribute
+ * which indicates the index of an item in {@link moonstone/VirtualList.VirtualList}
+ * or {@link moonstone/VirtualList.VirtualGridList}.
+ *
+ * @constant dataIndexAttribute
+ * @memberof moonstone/Scroller
+ * @type {String}
+ * @private
+ */
+const dataIndexAttribute = 'data-index';
+
+/**
+ * {@link moonstone/Scroller.ScrollableNative} is a Higher-order Component
  * that applies a Scrollable behavior to its wrapped component.
  *
  * Scrollable catches `onFocus` event from its wrapped component for spotlight features,
- * and also catches `onMouseDown`, `onMouseLeave`, `onMouseMove`, `onMouseUp`, `onWheel` and `onKeyDown` events
- * from its wrapped component for scrolling behaviors.
+ * and also catches `onWheel`, `onScroll` and `onKeyDown` events from its wrapped component for scrolling behaviors.
  *
  * Scrollable calls `onScrollStart`, `onScroll`, and `onScrollStop` callback functions during scroll.
  *
- * @class Scrollable
+ * @class ScrollableNative
  * @memberof moonstone/Scroller
  * @hoc
  * @private
  */
 const ScrollableHoC = hoc((config, Wrapped) => {
 	return class Scrollable extends Component {
-		static displayName = 'Scrollable'
+		static displayName = 'ScrollableNative'
 
-		static propTypes = /** @lends moonstone/Scroller.Scrollable.prototype */ {
+		static propTypes = /** @lends moonstone/Scroller.ScrollableNative.prototype */ {
 			/**
 			 * The callback function which is called for linking scrollTo function.
 			 * You should specify a callback function as the value of this prop
@@ -164,7 +161,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			onScroll: PropTypes.func,
 
 			/**
-			 * Called when scrollbar visibility changes
+			 * Called when scrollbar visability changes
 			 *
 			 * @type {Function}
 			 * @public
@@ -216,23 +213,18 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			verticalScrollbar: 'auto'
 		}
 
-		static childContextTypes = {
-			...contextTypesResize,
-			...contextTypesState
-		}
-
-		static contextTypes = contextTypesState
+		static childContextTypes = contextTypesResize
+		static contextTypes = contextTypesRtl
 
 		constructor (props) {
 			super(props);
 
 			this.state = {
-				rtl: false,
-				remeasure: false,
 				isHorizontalScrollbarVisible: props.horizontalScrollbar === 'visible',
 				isVerticalScrollbarVisible: props.verticalScrollbar === 'visible'
 			};
 
+			this.initInnerChildRef = this.initRef('innerChildRef');
 			this.initChildRef = this.initRef('childRef');
 			this.initContainerRef = this.initRef('containerRef');
 
@@ -259,28 +251,15 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 
 		getChildContext () {
 			return {
-				invalidateBounds: this.enqueueForceUpdate,
-				Subscriber: this.publisher.getSubscriber()
+				invalidateBounds: this.enqueueForceUpdate
 			};
-		}
-
-		componentWillMount () {
-			this.publisher = Publisher.create('resize', this.context.Subscriber);
-			this.publisher.publish({
-				remeasure: false
-			});
-
-			if (this.context.Subscriber) {
-				this.context.Subscriber.subscribe('resize', this.handleSubscription);
-				this.context.Subscriber.subscribe('i18n', this.handleSubscription);
-			}
 		}
 
 		componentDidMount () {
 			const bounds = this.getScrollBounds();
 
 			this.pageDistance = (this.canScrollVertically(bounds) ? bounds.clientHeight : bounds.clientWidth) * paginationPageMultiplier;
-			this.direction = this.childRef.props.direction;
+			this.direction = this.innerChildRef.props.direction;
 			this.updateEventListeners();
 			this.updateScrollbars();
 
@@ -291,23 +270,21 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			this.deferScrollTo = true;
 		}
 
-		componentDidUpdate (prevProps, prevState) {
+		componentDidUpdate () {
 			// Need to sync calculated client size if it is different from the real size
-			if (this.childRef.syncClientSize) {
+			if (this.innerChildRef.syncClientSize) {
 				const {isVerticalScrollbarVisible, isHorizontalScrollbarVisible} = this.state;
 				// If we actually synced, we need to reset scroll position.
-				if (this.childRef.syncClientSize()) {
+				if (this.innerChildRef.syncClientSize()) {
 					this.setScrollLeft(0);
 					this.setScrollTop(0);
 				}
 				// Need to check item total size is same with client size when scrollbar is visible
 				// By hiding scrollbar again, infinite function call maybe happens
-				this.isFitClientSize = (isVerticalScrollbarVisible || isHorizontalScrollbarVisible) && this.childRef.isSameTotalItemSizeWithClient();
+				this.isFitClientSize = (isVerticalScrollbarVisible || isHorizontalScrollbarVisible) && this.innerChildRef.isSameTotalItemSizeWithClient();
 			}
 
-			this.clampScrollPosition();
-
-			this.direction = this.childRef.props.direction;
+			this.direction = this.innerChildRef.props.direction;
 			this.updateEventListeners();
 			if (!this.isFitClientSize) {
 				this.updateScrollbars();
@@ -320,78 +297,50 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			} else {
 				this.updateScrollOnFocus();
 			}
-
-			// publish container resize changes
-			const horizontal = this.state.isHorizontalScrollbarVisible !== prevState.isHorizontalScrollbarVisible;
-			const vertical = this.state.isVerticalScrollbarVisible !== prevState.isVerticalScrollbarVisible;
-			if (horizontal || vertical) {
-				this.publisher.publish({
-					horizontal,
-					vertical
-				});
-			}
 		}
 
 		componentWillUnmount () {
 			const
 				{containerRef} = this,
-				childContainerRef = this.childRef.containerRef;
+				childContainerRef = this.innerChildRef.containerRef;
 
 			// Before call cancelAnimationFrame, you must send scrollStop Event.
-			if (this.animator.isAnimating()) {
+			if (this.scrolling) {
 				this.doScrollStop();
-				this.animator.stop();
 			}
+			this.scrollStopJob.stop();
 
 			if (containerRef && containerRef.removeEventListener) {
 				// FIXME `onWheel` doesn't work on the v8 snapshot.
 				containerRef.removeEventListener('wheel', this.onWheel);
 			}
 			if (childContainerRef && childContainerRef.removeEventListener) {
+				// FIXME `onScroll` doesn't work on the v8 snapshot.
+				childContainerRef.removeEventListener('scroll', this.onScroll, {capture: true});
 				// FIXME `onFocus` doesn't work on the v8 snapshot.
 				childContainerRef.removeEventListener('focusin', this.onFocus);
+				// FIXME `onMouseOver` doesn't work on the v8 snapshot.
+				childContainerRef.removeEventListener('mouseover', this.onMouseOver, {capture: true});
+				// FIXME `onMouseMove` doesn't work on the v8 snapshot.
+				childContainerRef.removeEventListener('mousemove', this.onMouseMove, {capture: true});
 			}
+
 			off('keydown', this.onKeyDown);
-
-			if (this.context.Subscriber) {
-				this.context.Subscriber.unsubscribe('resize', this.handleSubscription);
-				this.context.Subscriber.unsubscribe('i18n', this.handleSubscription);
-			}
-		}
-
-		handleSubscription = ({channel, message}) => {
-			if (channel === 'i18n') {
-				const {rtl} = message;
-				if (rtl !== this.state.rtl) {
-					this.setState({rtl});
-				}
-			} else if (channel === 'resize') {
-				this.publisher.publish(message);
-			}
 		}
 
 		// status
 		direction = 'vertical'
 		isScrollAnimationTargetAccumulated = false
-		wheelDirection = 0
-		pageDirection = 0
-		isFirstDragging = false
-		isDragging = false
 		deferScrollTo = true
 		pageDistance = 0
-		isWheeling = false
+		animateOnFocus = false
+		pageDirection = 0
 		isFitClientSize = false
+		isWheeling = false
 		isUpdatedScrollThumb = false
 
-		// drag info
-		dragInfo = {
-			t: 0,
-			clientX: 0,
-			clientY: 0,
-			dx: 0,
-			dy: 0,
-			dt: 0
-		}
+		// event handlers
+		eventHandlers = {}
 
 		// bounds info
 		bounds = {
@@ -416,80 +365,12 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		nodeToFocus = null
 
 		// component info
+		innerChildRef = null
 		childRef = null
 		containerRef = null
 
-		// scroll animator
-		animator = new ScrollAnimator()
-
-		// handle an input event
-
-		dragStart (e) {
-			const d = this.dragInfo;
-
-			this.isDragging = true;
-			this.isFirstDragging = true;
-			d.t = perfNow();
-			d.clientX = e.clientX;
-			d.clientY = e.clientY;
-			d.dx = d.dy = 0;
-		}
-
-		drag (e) {
-			const
-				{direction} = this,
-				t = perfNow(),
-				d = this.dragInfo;
-
-			if (direction === 'horizontal' || direction === 'both') {
-				d.dx = e.clientX - d.clientX;
-				d.clientX = e.clientX;
-			} else {
-				d.dx = 0;
-			}
-
-			if (direction === 'vertical' || direction === 'both') {
-				d.dy = e.clientY - d.clientY;
-				d.clientY = e.clientY;
-			} else {
-				d.dy = 0;
-			}
-
-			d.t = t;
-
-			return {dx: d.dx, dy: d.dy};
-		}
-
-		dragStop () {
-			const
-				d = this.dragInfo,
-				t = perfNow();
-
-			d.dt = t - d.t;
-			this.isDragging = false;
-		}
-
-		isFlicking () {
-			const d = this.dragInfo;
-
-			if (d.dt > holdTime) {
-				return false;
-			} else {
-				return true;
-			}
-		}
-
-		clampScrollPosition () {
-			const bounds = this.getScrollBounds();
-
-			if (this.scrollTop > bounds.maxTop) {
-				this.scrollTop = bounds.maxTop;
-			}
-
-			if (this.scrollLeft > bounds.maxLeft) {
-				this.scrollLeft = bounds.maxLeft;
-			}
-		}
+		// browser native scrolling
+		resetPosition = null // prevent auto-scroll on focus by Spotlight
 
 		calculateDistanceByWheel (deltaMode, delta, maxPixel) {
 			if (deltaMode === 0) {
@@ -503,108 +384,124 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			return delta;
 		}
 
-		// mouse event handler for JS scroller
+		// event handler for browser native scroll
 
-		onMouseDown = (e) => {
-			this.animator.stop();
-			this.dragStart(e);
+		onMouseDown = () => {
+			this.isScrollAnimationTargetAccumulated = false;
+			this.lastFocusedItem = null;
+			this.childRef.setContainerDisabled(false);
 		}
 
-		onMouseMove = (e) => {
-			if (this.isDragging) {
-				const
-					{dx, dy} = this.drag(e),
-					bounds = this.getScrollBounds();
+		onMouseOver = () => {
+			this.resetPosition = this.innerChildRef.containerRef.scrollTop;
+		}
 
-				if (this.isFirstDragging) {
-					if (!this.scrolling) {
-						this.scrolling = true;
-						this.doScrollStart();
-					}
-					this.isFirstDragging = false;
-				}
-				this.showThumb(bounds);
-				this.scroll(this.scrollLeft - dx, this.scrollTop - dy);
+		onMouseMove = () => {
+			if (this.resetPosition !== null) {
+				const childContainerRef = this.innerChildRef.containerRef;
+				childContainerRef.style.scrollBehavior = null;
+				childContainerRef.scrollTop = this.resetPosition;
+				childContainerRef.style.scrollBehavior = 'smooth';
+				this.resetPosition = null;
 			}
 		}
 
-		onMouseUp = (e) => {
-			if (this.isDragging) {
-				this.dragStop(e);
-
-				if (!this.isFlicking()) {
-					this.stop();
-				} else {
-					const
-						d = this.dragInfo,
-						target = this.animator.simulate(
-							this.scrollLeft,
-							this.scrollTop,
-							calcVelocity(-d.dx, d.dt),
-							calcVelocity(-d.dy, d.dt)
-						),
-						focusedItem = Spotlight.getCurrent();
-
-					if (focusedItem) {
-						focusedItem.blur();
-					}
-					this.childRef.setContainerDisabled(true);
-					this.isScrollAnimationTargetAccumulated = false;
-					this.start({
-						targetX: target.targetX,
-						targetY: target.targetY,
-						animate: true,
-						duration: target.duration
-					});
-				}
-			}
-		}
-
-		onMouseLeave = (e) => {
-			this.onMouseMove(e);
-			this.onMouseUp();
-		}
-
+		/*
+		 * wheel event handler;
+		 * - for horizontal scroll, supports wheel action on any children nodes since web engine cannot suppor this
+		 * - for vertical scroll, supports wheel action on scrollbars only
+		 */
 		onWheel = (e) => {
-			e.preventDefault();
-			if (!this.isDragging) {
-				const
-					bounds = this.getScrollBounds(),
-					canScrollHorizontally = this.canScrollHorizontally(bounds),
-					canScrollVertically = this.canScrollVertically(bounds),
-					focusedItem = Spotlight.getCurrent(),
-					eventDeltaMode = e.deltaMode,
-					eventDelta = (-e.wheelDeltaY || e.deltaY),
-					isVerticalScrollButtonFocused = this.verticalScrollbarRef && this.verticalScrollbarRef.isThumbFocused(),
-					isHorizontalScrollButtonFocused = this.horizontalScrollbarRef && this.horizontalScrollbarRef.isThumbFocused();
-				let
-					delta = 0,
-					direction;
+			const
+				bounds = this.getScrollBounds(),
+				canScrollHorizontally = this.canScrollHorizontally(bounds),
+				canScrollVertically = this.canScrollVertically(bounds),
+				eventDeltaMode = e.deltaMode,
+				eventDelta = (-e.wheelDeltaY || e.deltaY);
+			let
+				delta = 0,
+				needToHideThumb = false;
 
-				if (canScrollVertically) {
-					delta = this.calculateDistanceByWheel(eventDeltaMode, eventDelta, bounds.clientHeight * scrollWheelPageMultiplierForMaxPixel);
-				} else if (canScrollHorizontally) {
+			this.lastFocusedItem = null;
+			if (typeof window !== 'undefined') {
+				window.document.activeElement.blur();
+			}
+
+			this.showThumb(bounds);
+
+			// FIXME This routine is a temporary support for horizontal wheel scroll.
+			// FIXME If web engine supports horizontal wheel, this routine should be refined or removed.
+			if (canScrollVertically) { // This routine handles wheel events on scrollbars for vertical scroll.
+				if (eventDelta < 0 && this.scrollTop > 0 || eventDelta > 0 && this.scrollTop < bounds.maxTop) {
+					const {horizontalScrollbarRef, verticalScrollbarRef} = this;
+
+					if (!this.isWheeling) {
+						this.childRef.setContainerDisabled(true);
+						this.isWheeling = true;
+					}
+
+					// Not to check if e.target is a descendant of a wrapped component which may have a lot of nodes in it.
+					if ((horizontalScrollbarRef && horizontalScrollbarRef.containerRef.contains(e.target)) ||
+						(verticalScrollbarRef && verticalScrollbarRef.containerRef.contains(e.target))) {
+						delta = this.calculateDistanceByWheel(eventDeltaMode, eventDelta, bounds.clientHeight * scrollWheelPageMultiplierForMaxPixel);
+						needToHideThumb = !delta;
+					}
+				} else {
+					needToHideThumb = true;
+				}
+			} else if (canScrollHorizontally) { // this routine handles wheel events on any children for horizontal scroll.
+				if (eventDelta < 0 && this.scrollLeft > 0 || eventDelta > 0 && this.scrollLeft < bounds.maxLeft) {
+					if (!this.isWheeling) {
+						this.childRef.setContainerDisabled(true);
+						this.isWheeling = true;
+					}
 					delta = this.calculateDistanceByWheel(eventDeltaMode, eventDelta, bounds.clientWidth * scrollWheelPageMultiplierForMaxPixel);
-				}
-
-				direction = Math.sign(delta);
-
-				if (focusedItem && !isVerticalScrollButtonFocused && !isHorizontalScrollButtonFocused) {
-					focusedItem.blur();
-				}
-
-				if (direction !== this.wheelDirection) {
-					this.isScrollAnimationTargetAccumulated = false;
-					this.wheelDirection = direction;
-				}
-
-				if (delta !== 0) {
-					this.isWheeling = true;
-					this.childRef.setContainerDisabled(true);
-					this.scrollToAccumulatedTarget(delta, canScrollVertically);
+					needToHideThumb = !delta;
+				} else {
+					needToHideThumb = true;
 				}
 			}
+
+			if (delta !== 0) {
+				/* prevent native scrolling feature for vertical direction */
+				e.preventDefault();
+				const direction = Math.sign(delta);
+				// Not to accumulate scroll position if wheel direction is different from hold direction
+				if (direction !== this.pageDirection) {
+					this.isScrollAnimationTargetAccumulated = false;
+					this.pageDirection = direction;
+				}
+				this.scrollToAccumulatedTarget(delta, canScrollVertically);
+			}
+
+			if (needToHideThumb) {
+				this.startHidingThumb();
+			}
 		}
+
+		onScroll = (e) => {
+			const
+				bounds = this.getScrollBounds(),
+				canScrollHorizontally = this.canScrollHorizontally(bounds);
+			let
+				{scrollLeft, scrollTop} = e.target;
+
+			if (!this.scrolling) {
+				this.scrollStartOnScroll();
+			}
+
+			if (this.context.rtl && canScrollHorizontally) {
+				/* FIXME: RTL / this calculation only works for Chrome */
+				scrollLeft = bounds.maxLeft - scrollLeft;
+			}
+
+			this.scroll(scrollLeft, scrollTop);
+
+			this.startHidingThumb();
+			this.scrollStopJob.start();
+		}
+
+		// event handlers for Spotlight support
 
 		startScrollOnFocus = (pos, item) => {
 			if (pos) {
@@ -613,12 +510,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					bounds = this.getScrollBounds();
 
 				if ((bounds.maxTop > 0 && top !== this.scrollTop) || (bounds.maxLeft > 0 && left !== this.scrollLeft)) {
-					this.start({
-						targetX: left,
-						targetY: top,
-						animate: (animationDuration > 0) && this.animateOnFocus,
-						duration: animationDuration
-					});
+					this.start(left, top, this.animateOnFocus);
 				}
 				this.lastFocusedItem = item;
 				this.lastScrollPositionOnFocus = pos;
@@ -630,28 +522,29 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				this.childRef.shouldPreventScrollByFocus() :
 				false;
 
-			if (this.isWheeling) {
-				this.stop();
-				this.animateOnFocus = false;
-			}
-
 			if (!Spotlight.getPointerMode()) {
 				this.alertThumb();
 			}
 
-			if (!(shouldPreventScrollByFocus || Spotlight.getPointerMode() || this.isDragging)) {
+			if (!(shouldPreventScrollByFocus || Spotlight.getPointerMode())) {
 				const
 					item = e.target,
-					positionFn = this.childRef.calculatePositionOnFocus,
+					positionFn = this.innerChildRef.calculatePositionOnFocus,
 					spotItem = Spotlight.getCurrent();
 
 				if (item && item === spotItem && positionFn) {
 					const lastPos = this.lastScrollPositionOnFocus;
 					let pos;
 
+					const focusedIndex = Number.parseInt(item.getAttribute(dataIndexAttribute));
+					if (!isNaN(focusedIndex)) {
+						this.childRef.setNodeIndexToBeFocused(null);
+						this.childRef.setLastFocusedIndex(focusedIndex);
+					}
+
 					// If scroll animation is ongoing, we need to pass last target position to
 					// determine correct scroll position.
-					if (this.animator.isAnimating() && lastPos) {
+					if (this.scrolling && lastPos) {
 						pos = positionFn({item, scrollPosition: (this.direction !== 'horizontal') ? lastPos.top : lastPos.left});
 					} else {
 						pos = positionFn({item});
@@ -708,19 +601,23 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				{getEndPoint, scrollToAccumulatedTarget} = this,
 				bounds = this.getScrollBounds(),
 				canScrollVertically = this.canScrollVertically(bounds),
+				innerChildRef = this.innerChildRef,
 				childRef = this.childRef,
 				pageDistance = isPageUp(keyCode) ? (this.pageDistance * -1) : this.pageDistance,
 				spotItem = Spotlight.getCurrent();
 
 			if (!Spotlight.getPointerMode() && spotItem) {
 				// Should skip scroll by page when spotItem is paging control button of Scrollbar
-				if (!childRef.containerRef.contains(spotItem)) {
+				if (!innerChildRef.containerRef.contains(spotItem)) {
 					return;
 				}
-
 				const
-					// VirtualList and Scroller have a containerId on containerRef
-					containerId = childRef.containerRef.dataset.containerId,
+					containerId = (
+						// ScrollerNative has a containerId on containerRef
+						innerChildRef.containerRef.dataset.containerId ||
+						// VirtualListNative has a containerId on contentRef
+						innerChildRef.contentRef.dataset.containerId
+					),
 					direction = this.getPageDirection(keyCode),
 					rDirection = reverseDirections[direction],
 					viewportBounds = this.containerRef.getBoundingClientRect(),
@@ -767,8 +664,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 
 		onKeyDown = (e) => {
 			this.animateOnFocus = true;
-			if ((isPageUp(e.keyCode) || isPageDown(e.keyCode)) && !e.repeat && this.hasFocus()) {
-				this.scrollByPage(e.keyCode);
+			if (isPageUp(e.keyCode) || isPageDown(e.keyCode)) {
+				e.preventDefault();
+				if (!e.repeat && this.hasFocus()) {
+					this.scrollByPage(e.keyCode);
+				}
 			}
 		}
 
@@ -795,16 +695,12 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			}
 
 			if (vertical) {
-				this.accumulatedTargetY = this.accumulatedTargetY + delta;
+				this.accumulatedTargetY += delta;
 			} else {
-				this.accumulatedTargetX = this.accumulatedTargetX + delta;
+				this.accumulatedTargetX += delta;
 			}
 
-			this.start({
-				targetX: this.accumulatedTargetX,
-				targetY: this.accumulatedTargetY,
-				animate: true
-			});
+			this.start(this.accumulatedTargetX, this.accumulatedTargetY);
 		}
 
 		// call scroll callbacks
@@ -820,6 +716,28 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		doScrollStop () {
 			forwardScrollStop({scrollLeft: this.scrollLeft, scrollTop: this.scrollTop, moreInfo: this.getMoreInfo()}, this.props);
 		}
+
+		// call scroll callbacks and update scrollbars for native scroll
+
+		scrollStartOnScroll = () => {
+			this.scrolling = true;
+			this.showThumb(this.getScrollBounds());
+			this.doScrollStart();
+		}
+
+		scrollStopOnScroll = () => {
+			this.isScrollAnimationTargetAccumulated = false;
+			this.childRef.setContainerDisabled(false);
+			this.focusOnItem();
+			this.lastFocusedItem = null;
+			this.lastScrollPositionOnFocus = null;
+
+			this.scrolling = false;
+			this.doScrollStop();
+			this.isWheeling = false;
+		}
+
+		scrollStopJob = new Job(this.scrollStopOnScroll.bind(this), scrollStopWaiting);
 
 		// update scroll position
 
@@ -841,105 +759,62 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			}
 		}
 
-		// scroll start/stop
+		// scroll start
 
-		start ({targetX, targetY, animate = true, duration = animationDuration}) {
-			const {scrollLeft, scrollTop} = this;
-			const bounds = this.getScrollBounds();
+		start (targetX, targetY, animate = true) {
+			const
+				bounds = this.getScrollBounds(),
+				childContainerRef = this.innerChildRef.containerRef;
 
-			this.animator.stop();
-			if (!this.scrolling) {
-				this.scrolling = true;
-				this.doScrollStart();
-			}
+			targetX = clamp(0, bounds.maxLeft, targetX);
+			targetY = clamp(0, bounds.maxTop, targetY);
 
-			if (Math.abs(bounds.maxLeft - targetX) < epsilon) {
+			if ((bounds.maxLeft - targetX) < epsilon) {
 				targetX = bounds.maxLeft;
 			}
-			if (Math.abs(bounds.maxTop - targetY) < epsilon) {
+			if ((bounds.maxTop - targetY) < epsilon) {
 				targetY = bounds.maxTop;
 			}
 
-			this.showThumb(bounds);
-
 			if (animate) {
-				this.animator.animate(this.scrollAnimation({
-					sourceX: scrollLeft,
-					sourceY: scrollTop,
-					targetX,
-					targetY,
-					duration
-				}));
+				this.innerChildRef.scrollToPosition(targetX, targetY);
 			} else {
-				targetX = clamp(0, bounds.maxLeft, targetX);
-				targetY = clamp(0, bounds.maxTop, targetY);
-
-				this.scroll(targetX, targetY);
-				this.stop();
+				childContainerRef.style.scrollBehavior = null;
+				this.innerChildRef.scrollToPosition(targetX, targetY);
+				childContainerRef.style.scrollBehavior = 'smooth';
+				this.focusOnItem();
 			}
 		}
 
-		scrollAnimation = (animationInfo) => (curTime) => {
-			const
-				{sourceX, sourceY, targetX, targetY, duration} = animationInfo,
-				bounds = this.getScrollBounds();
-
-			if (curTime < duration) {
-				this.scroll(
-					this.canScrollHorizontally(bounds) ? clamp(0, bounds.maxLeft, this.animator.timingFunction(sourceX, targetX, duration, curTime)) : sourceX,
-					this.canScrollVertically(bounds) ? clamp(0, bounds.maxTop, this.animator.timingFunction(sourceY, targetY, duration, curTime)) : sourceY
-				);
-			} else {
-				this.scroll(
-					clamp(0, bounds.maxLeft, targetX),
-					clamp(0, bounds.maxTop, targetY)
-				);
-				this.stop();
+		focusOnItem () {
+			if (this.indexToFocus !== null && typeof this.innerChildRef.focusByIndex === 'function') {
+				this.innerChildRef.focusByIndex(this.indexToFocus);
+				this.indexToFocus = null;
+			}
+			if (this.nodeToFocus !== null && typeof this.innerChildRef.focusOnNode === 'function') {
+				this.innerChildRef.focusOnNode(this.nodeToFocus);
+				this.nodeToFocus = null;
 			}
 		}
 
 		scroll = (left, top) => {
 			let
-				dirX = 0,
-				dirY = 0;
+				dirHorizontal = 0,
+				dirVertical = 0;
 
 			if (left !== this.scrollLeft) {
-				dirX = Math.sign(left - this.scrollLeft);
+				dirHorizontal = Math.sign(left - this.scrollLeft);
 				this.setScrollLeft(left);
 			}
 			if (top !== this.scrollTop) {
-				dirY = Math.sign(top - this.scrollTop);
+				dirVertical = Math.sign(top - this.scrollTop);
 				this.setScrollTop(top);
 			}
 
-			this.childRef.setScrollPosition(this.scrollLeft, this.scrollTop, dirX, dirY);
+			if (this.innerChildRef.didScroll) {
+				this.innerChildRef.didScroll(this.scrollLeft, this.scrollTop, dirHorizontal, dirVertical);
+			}
 			this.doScrolling();
-		}
-
-		stop () {
-			this.animator.stop();
-			this.isScrollAnimationTargetAccumulated = false;
-			this.childRef.setContainerDisabled(false);
-			this.focusOnItem();
-			this.lastFocusedItem = null;
-			this.lastScrollPositionOnFocus = null;
-			this.startHidingThumb();
-			this.isWheeling = false;
-			if (this.scrolling) {
-				this.scrolling = false;
-				this.doScrollStop();
-			}
-		}
-
-		focusOnItem () {
-			if (this.indexToFocus !== null && typeof this.childRef.focusByIndex === 'function') {
-				this.childRef.focusByIndex(this.indexToFocus);
-				this.indexToFocus = null;
-			}
-			if (this.nodeToFocus !== null && typeof this.childRef.focusOnNode === 'function') {
-				this.childRef.focusOnNode(this.nodeToFocus);
-				this.nodeToFocus = null;
-			}
 		}
 
 		// scrollTo API
@@ -984,11 +859,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 						}
 					}
 				} else {
-					if (typeof opt.index === 'number' && typeof this.childRef.getItemPosition === 'function') {
-						itemPos = this.childRef.getItemPosition(opt.index, opt.stickTo);
+					if (typeof opt.index === 'number' && typeof this.innerChildRef.getItemPosition === 'function') {
+						itemPos = this.innerChildRef.getItemPosition(opt.index, opt.stickTo);
 					} else if (opt.node instanceof Object) {
-						if (opt.node.nodeType === 1 && typeof this.childRef.getNodePosition === 'function') {
-							itemPos = this.childRef.getNodePosition(opt.node);
+						if (opt.node.nodeType === 1 && typeof this.innerChildRef.getNodePosition === 'function') {
+							itemPos = this.innerChildRef.getNodePosition(opt.node);
 						}
 					}
 					if (itemPos) {
@@ -1011,11 +886,11 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				this.indexToFocus = (opt.focus && typeof opt.index === 'number') ? opt.index : null;
 				this.nodeToFocus = (opt.focus && opt.node instanceof Object && opt.node.nodeType === 1) ? opt.node : null;
 				this.scrollToInfo = null;
-				this.start({
-					targetX: (left !== null) ? left : this.scrollLeft,
-					targetY: (top !== null) ? top : this.scrollTop,
-					animate: opt.animate
-				});
+				this.start(
+					(left !== null) ? left : this.scrollLeft,
+					(top !== null) ? top : this.scrollTop,
+					opt.animate
+				);
 			} else {
 				this.scrollToInfo = opt;
 			}
@@ -1087,7 +962,6 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				if (this.props.onScrollbarVisibilityChange) {
 					this.props.onScrollbarVisibilityChange();
 				}
-
 				// one or both scrollbars have changed visibility
 				this.setState({
 					isHorizontalScrollbarVisible: curHorizontalScrollbarVisible,
@@ -1118,10 +992,10 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 						scrollTop: this.scrollTop
 					};
 
-				if (curHorizontalScrollbarVisible && this.horizontalScrollbarRef) {
+				if (curHorizontalScrollbarVisible) {
 					this.horizontalScrollbarRef.update(updatedBounds);
 				}
-				if (curVerticalScrollbarVisible && this.verticalScrollbarRef) {
+				if (curVerticalScrollbarVisible) {
 					this.verticalScrollbarRef.update(updatedBounds);
 				}
 				return true;
@@ -1132,43 +1006,52 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		alertThumbAfterRendered = () => {
 			const spotItem = Spotlight.getCurrent();
 
-			if (!Spotlight.getPointerMode() && spotItem && this.childRef.containerRef.contains(spotItem) && this.isUpdatedScrollThumb) {
+			if (!Spotlight.getPointerMode() && spotItem && this.innerChildRef.containerRef.contains(spotItem) && this.isUpdatedScrollThumb) {
 				this.alertThumb();
 			}
 		}
+
 		// ref
 
 		getScrollBounds () {
-			if (typeof this.childRef.getScrollBounds === 'function') {
-				return this.childRef.getScrollBounds();
+			if (typeof this.innerChildRef.getScrollBounds === 'function') {
+				return this.innerChildRef.getScrollBounds();
 			}
 		}
 
 		getMoreInfo () {
-			if (typeof this.childRef.getMoreInfo === 'function') {
-				return this.childRef.getMoreInfo();
+			if (typeof this.innerChildRef.getMoreInfo === 'function') {
+				return this.innerChildRef.getMoreInfo();
 			}
 		}
 
 		updateEventListeners = () => {
 			const
 				{containerRef} = this,
-				childContainerRef = this.childRef.containerRef;
+				childContainerRef = this.innerChildRef.containerRef;
 
 			if (containerRef && containerRef.addEventListener) {
 				// FIXME `onWheel` doesn't work on the v8 snapshot.
 				containerRef.addEventListener('wheel', this.onWheel);
 			}
 			if (childContainerRef && childContainerRef.addEventListener) {
+				// FIXME `onScroll` doesn't work on the v8 snapshot.
+				childContainerRef.addEventListener('scroll', this.onScroll, {capture: true});
 				// FIXME `onFocus` doesn't work on the v8 snapshot.
 				childContainerRef.addEventListener('focusin', this.onFocus);
+				// FIXME `onMouseOver` doesn't work on the v8 snapshot.
+				childContainerRef.addEventListener('mouseover', this.onMouseOver, {capture: true});
+				// FIXME `onMouseMove` doesn't work on the v8 snapshot.
+				childContainerRef.addEventListener('mousemove', this.onMouseMove, {capture: true});
 			}
+
+			childContainerRef.style.scrollBehavior = 'smooth';
 		}
 
 		updateScrollOnFocus () {
 			const
 				focusedItem = Spotlight.getCurrent(),
-				{containerRef, calculatePositionOnFocus} = this.childRef;
+				{containerRef, calculatePositionOnFocus} = this.innerChildRef;
 
 			if (focusedItem && containerRef && containerRef.contains(focusedItem)) {
 				const
@@ -1179,11 +1062,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					pos = calculatePositionOnFocus({item: focusedItem, scrollInfo});
 
 				if (pos && (pos.left !== this.scrollLeft || pos.top !== this.scrollTop)) {
-					this.start({
-						targetX: pos.left,
-						targetY: pos.top,
-						animate: false
-					});
+					this.start(pos.left, pos.top, false);
 				}
 			}
 
@@ -1194,7 +1073,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		// TODO: consider replacing forceUpdate() by storing bounds in state rather than a non-
 		// state member.
 		enqueueForceUpdate = () => {
-			this.childRef.calculateMetrics();
+			this.innerChildRef.calculateMetrics();
 			this.forceUpdate();
 		}
 
@@ -1204,13 +1083,6 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			return (ref) => {
 				this[prop] = ref;
 			};
-		}
-
-		handleScroll = () => {
-			if (!this.animator.isAnimating() && this.childRef && this.childRef.containerRef) {
-				this.childRef.containerRef.scrollTop = this.scrollTop;
-				this.childRef.containerRef.scrollLeft = this.scrollLeft;
-			}
 		}
 
 		render () {
@@ -1232,7 +1104,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			delete props.verticalScrollbar;
 
 			return (
-				<ScrollableSpotlightContainer
+				<ScrollableSpotlightContainerDecorator
 					className={scrollableClasses}
 					containerRef={this.initContainerRef}
 					focusableScrollbar={focusableScrollbar}
@@ -1243,17 +1115,17 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 							{...props}
 							cbScrollTo={this.scrollTo}
 							className={css.content}
-							onScroll={this.handleScroll}
+							innerRef={this.initInnerChildRef}
 							ref={this.initChildRef}
 						/>
 						{isVerticalScrollbarVisible ? <Scrollbar {...this.verticalScrollbarProps} disabled={!isVerticalScrollbarVisible} /> : null}
 					</div>
 					{isHorizontalScrollbarVisible ? <Scrollbar {...this.horizontalScrollbarProps} corner={isVerticalScrollbarVisible} disabled={!isHorizontalScrollbarVisible} /> : null}
-				</ScrollableSpotlightContainer>
+				</ScrollableSpotlightContainerDecorator>
 			);
 		}
 	};
 });
 
 export default ScrollableHoC;
-export {dataIndexAttribute, ScrollableHoC as Scrollable};
+export {dataIndexAttribute, ScrollableHoC as ScrollableNative};
