+++ packages/ui/Scrollable/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Scrollable.js"
+}
