+++ packages/ui/Scrollbar/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Scrollbar.js"
+}
