--- packages/ui/Scroller/Scrollable.less
@@ -1,27 +1,0 @@
-.scrollableFill {
-	position: relative;
-	width: 100%;
-	height: 100%;
-}
-
-.scrollable {
-	.scrollableFill;
-
-	display: flex;
-	box-sizing: border-box;
-	flex-direction: column;
-
-	.container {
-		display: flex;
-		width: 100%;
-		height: 100%;
-		flex: auto;
-	}
-
-	.content {
-		background: inherit;
-		position: relative;
-		overflow: hidden;
-		flex: auto;
-	}
-}
