--- packages/ui/Scroller/Scroller.less
@@ -1,5 +1,0 @@
-.hideNativeScrollbar {
-	&::-webkit-scrollbar {
-		display: none;
-	}
-}
