--- packages/ui/VirtualList/ListItem.less
@@ -1,7 +1,0 @@
-.listItem {
-	position: absolute;
-	overflow: hidden;
-	width: 100%;
-
-	will-change: transform;
-}
