+++ packages/ui/VirtualList/VirtualListBaseNative.js
@@ -1,35 +1,22 @@
 /*
- * Exports the {@link moonstone/VirtualList.VirtualListBase} and
- * {@link moonstone/VirtualList.VirtualListCore} components and the
+ * Exports the {@link moonstone/VirtualList.VirtualListBaseNative} and
+ * {@link moonstone/VirtualList.VirtualListCoreNative} components and the
  * {@link moonstone/VirtualList.gridListItemSizeShape} validator. The default
- * export is {@link moonstone/VirtualList.VirtualListBase}.
+ * export is {@link moonstone/VirtualList.VirtualListBaseNative}.
  */
 
-import clamp from 'ramda/src/clamp';
 import classNames from 'classnames';
 import {contextTypes} from '@enact/i18n/I18nDecorator';
-import {forward} from '@enact/core/handle';
-import {is} from '@enact/core/keymap';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
-import Spotlight, {getDirection} from '@enact/spotlight';
-import Spottable from '@enact/spotlight/Spottable';
-import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 
-import {dataIndexAttribute, Scrollable} from '../Scroller/Scrollable';
+import {dataIndexAttribute, ScrollableNative} from '../Scrollable/ScrollableNative.js';
+import {VirtualListCoreSpottable, VirtualListContainerSpottable} from './VirtualListCoreSpottable';
 
-import css from './ListItem.less';
+import css from '@enact/moonstone/VirtualList/VirtualListNative.less';
+import cssItem from '@enact/moonstone/VirtualList/ListItem.less';
 
-const SpotlightPlaceholder = Spottable('div');
-
-const
-	dataContainerDisabledAttribute = 'data-container-disabled',
-	forwardKeyDown = forward('onKeyDown'),
-	nop = () => {},
-	isDown = is('down'),
-	isLeft = is('left'),
-	isRight = is('right'),
-	isUp = is('up');
+const nop = () => {};
 
 /**
  * The shape for the grid list item size in a list for {@link moonstone/VirtualList.listItemSizeShape}.
@@ -38,6 +25,7 @@ const
  * @memberof moonstone/VirtualList
  * @property {Number} minWidth - The minimum width of the grid list item.
  * @property {Number} minHeight - The minimum height of the grid list item.
+ * @private
  */
 const gridListItemSizeShape = PropTypes.shape({
 	minWidth: PropTypes.number.isRequired,
@@ -45,19 +33,19 @@ const gridListItemSizeShape = PropTypes.shape({
 });
 
 /**
- * {@link moonstone/VirtualList.VirtualListBase} is a base component for
+ * {@link moonstone/VirtualList.VirtualListBaseNative} is a base component for
  * {@link moonstone/VirtualList.VirtualList} and
  * {@link moonstone/VirtualList.VirtualGridList} with Scrollable and SpotlightContainerDecorator applied.
  *
- * @class VirtualListCore
+ * @class VirtualListCoreNative
  * @memberof moonstone/VirtualList
  * @ui
  * @private
  */
-class VirtualListCore extends Component {
-	static displayName = 'VirtualListBase'
+class VirtualListCoreNative extends Component {
+	static displayName = 'VirtualListBaseNative'
 
-	static propTypes = /** @lends moonstone/VirtualList.VirtualListCore.prototype */ {
+	static propTypes = /** @lends moonstone/VirtualList.VirtualListCoreNative.prototype */ {
 		/**
 		 * The render function for an item of the list.
 		 * `index` is for accessing the index of the item.
@@ -139,6 +127,12 @@ class VirtualListCore extends Component {
 		 */
 		direction: PropTypes.oneOf(['horizontal', 'vertical']),
 
+		initItemRef: PropTypes.func,
+
+		getNodeIndexToBeFocused: PropTypes.func,
+
+		lastFocusedIndex: PropTypes.object,
+
 		/**
 		 * Number of spare DOM node.
 		 * `3` is good for the default value experimentally and
@@ -160,13 +154,33 @@ class VirtualListCore extends Component {
 		pageScroll: PropTypes.bool,
 
 		/**
+		 * The index to restore
+		 *
+		 * @type {Number}
+		 * @default null
+		 * @public
+		 */
+		preservedIndex: PropTypes.object,
+
+		/**
+		 * The value whether restoring focus or not
+		 *
+		 * @type {Number}
+		 * @default null
+		 * @public
+		 */
+		restoreLastFocused: PropTypes.bool,
+
+		/**
 		 * Spacing between items.
 		 *
 		 * @type {Number}
 		 * @default 0
 		 * @public
 		 */
-		spacing: PropTypes.number
+		spacing: PropTypes.number,
+
+		withPlaceholder: PropTypes.func
 	}
 
 	static contextTypes = contextTypes
@@ -176,8 +190,12 @@ class VirtualListCore extends Component {
 		data: [],
 		dataSize: 0,
 		direction: 'vertical',
+		getNodeIndexToBeFocused: nop,
+		lastFocusedIndex: null,
 		overhang: 3,
 		pageScroll: false,
+		preservedIndex: null,
+		restoreLastFocused: false,
 		spacing: 0
 	}
 
@@ -185,8 +203,8 @@ class VirtualListCore extends Component {
 		super(props);
 
 		this.state = {firstIndex: 0, numOfItems: 0};
+		this.initContentRef = this.initRef('contentRef');
 		this.initContainerRef = this.initRef('containerRef');
-		this.initItemContainerRef = this.initRef('itemContainerRef');
 	}
 
 	componentWillMount () {
@@ -199,22 +217,11 @@ class VirtualListCore extends Component {
 	// Calculate metrics for VirtualList after the 1st render to know client W/H.
 	// We separate code related with data due to re use it when data changed.
 	componentDidMount () {
-		const containerNode = this.containerRef;
-
 		if (!this.props.clientSize) {
 			this.calculateMetrics(this.props);
 			this.updateStatesAndBounds(this.props);
 		}
-
-		// prevent native scrolling by Spotlight
-		this.preventScroll = () => {
-			containerNode.scrollTop = 0;
-			containerNode.scrollLeft = this.context.rtl ? containerNode.scrollWidth : 0;
-		};
-
-		if (containerNode && containerNode.addEventListener) {
-			containerNode.addEventListener('scroll', this.preventScroll);
-		}
+		this.setContainerSize();
 	}
 
 	// Call updateStatesAndBounds here when dataSize has been changed to update nomOfItems state.
@@ -233,24 +240,11 @@ class VirtualListCore extends Component {
 		if (hasMetricsChanged) {
 			this.calculateMetrics(nextProps);
 			this.updateStatesAndBounds(nextProps);
+			this.setContainerSize();
 		} else if (hasDataChanged) {
 			this.updateStatesAndBounds(nextProps);
+			this.setContainerSize();
 		}
-	}
-
-	componentDidUpdate () {
-		this.restoreFocus();
-	}
-
-	componentWillUnmount () {
-		const containerNode = this.containerRef;
-
-		// remove a function for preventing native scrolling by Spotlight
-		if (containerNode && containerNode.removeEventListener) {
-			containerNode.removeEventListener('scroll', this.preventScroll);
-		}
-
-		this.setContainerDisabled(false);
 	}
 
 	scrollBounds = {
@@ -276,74 +270,20 @@ class VirtualListCore extends Component {
 	dimensionToExtent = 0
 	threshold = 0
 	maxFirstIndex = 0
+	lastFirstIndex = 0
 	curDataSize = 0
 	cc = []
 	scrollPosition = 0
-	updateFrom = null
-	updateTo = null
-	isScrolledBy5way = false
+	isScrolledByJump = false
 
+	containerClass = null
+	contentRef = null
 	containerRef = null
-	itemContainerRef = null
-
-	// spotlight
-	nodeIndexToBeFocused = null
-	lastFocusedIndex = null
-	preservedIndex = null
 
 	isVertical = () => this.isPrimaryDirectionVertical
 
 	isHorizontal = () => !this.isPrimaryDirectionVertical
 
-	isPlaceholderFocused () {
-		const current = Spotlight.getCurrent();
-		if (current && current.dataset.vlPlaceholder && this.containerRef.contains(current)) {
-			return true;
-		}
-
-		return false;
-	}
-
-	handlePlaceholderFocus = (ev) => {
-		const placeholder = ev.currentTarget;
-
-		if (placeholder) {
-			const index = placeholder.dataset.index;
-
-			if (index) {
-				this.preservedIndex = parseInt(index);
-				this.restoreLastFocused = true;
-			}
-		}
-	}
-
-	restoreFocus () {
-		if (
-			this.restoreLastFocused &&
-			!this.isPlaceholderFocused()
-		) {
-			const containerId = this.props['data-container-id'];
-			const node = this.containerRef.querySelector(
-				`[data-container-id="${containerId}"] [data-index="${this.preservedIndex}"]`
-			);
-
-			if (node) {
-				// if we're supposed to restore focus and virtual list has positioned a set of items
-				// that includes lastFocusedIndex, clear the indicator
-				this.restoreLastFocused = false;
-
-				// try to focus the last focused item
-				const foundLastFocused = Spotlight.focus(node);
-
-				// but if that fails (because it isn't found or is disabled), focus the container so
-				// spotlight isn't lost
-				if (!foundLastFocused) {
-					this.restoreLastFocused = true;
-					Spotlight.focus(containerId);
-				}
-			}
-		}
-	}
 
 	getScrollBounds = () => this.scrollBounds
 
@@ -446,9 +386,9 @@ class VirtualListCore extends Component {
 
 	updateStatesAndBounds (props) {
 		const
-			{dataSize, overhang} = props,
+			{dataSize, overhang, preservedIndex, restoreLastFocused} = props,
 			{firstIndex} = this.state,
-			{dimensionToExtent, primary, moreInfo, preservedIndex, scrollPosition} = this,
+			{dimensionToExtent, primary, moreInfo, scrollPosition} = this,
 			numOfItems = Math.min(dataSize, dimensionToExtent * (Math.ceil(primary.clientSize / primary.gridSize) + overhang)),
 			wasFirstIndexMax = ((this.maxFirstIndex < moreInfo.firstVisibleIndex - dimensionToExtent) && (firstIndex === this.maxFirstIndex)),
 			dataSizeDiff = dataSize - this.curDataSize;
@@ -456,15 +396,13 @@ class VirtualListCore extends Component {
 
 		this.maxFirstIndex = Math.ceil((dataSize - numOfItems) / dimensionToExtent) * dimensionToExtent;
 		this.curDataSize = dataSize;
-		this.updateFrom = null;
-		this.updateTo = null;
 
 		// reset children
 		this.cc = [];
 		this.calculateScrollBounds(props);
 		this.updateMoreInfo(dataSize, scrollPosition);
 
-		if (this.restoreLastFocused &&
+		if (restoreLastFocused &&
 			numOfItems > 0 &&
 			(preservedIndex < dataSize) &&
 			(preservedIndex < moreInfo.firstVisibleIndex || preservedIndex > moreInfo.lastVisibleIndex)) {
@@ -543,19 +481,14 @@ class VirtualListCore extends Component {
 		if (this.scrollPosition > maxPos) {
 			this.props.cbScrollTo({position: (isPrimaryDirectionVertical) ? {y: maxPos} : {x: maxPos}, animate: false});
 		}
+
+		this.containerClass = (isPrimaryDirectionVertical) ? css.vertical : css.horizontal;
 	}
 
-	updateMoreInfo (dataSize, primaryPosition) {
-		const
-			{dimensionToExtent, moreInfo} = this,
-			{itemSize, gridSize, clientSize} = this.primary;
-
-		if (dataSize <= 0) {
-			moreInfo.firstVisibleIndex = null;
-			moreInfo.lastVisibleIndex = null;
-		} else {
-			moreInfo.firstVisibleIndex = (Math.floor((primaryPosition - itemSize) / gridSize) + 1) * dimensionToExtent;
-			moreInfo.lastVisibleIndex = Math.min(dataSize - 1, Math.ceil((primaryPosition + clientSize) / gridSize) * dimensionToExtent - 1);
+	setContainerSize = () => {
+		if (this.contentRef) {
+			this.contentRef.style.width = this.scrollBounds.scrollWidth + 'px';
+			this.contentRef.style.height = this.scrollBounds.scrollHeight + 'px';
 		}
 	}
 
@@ -573,10 +506,10 @@ class VirtualListCore extends Component {
 		}
 	}
 
-	setScrollPosition (x, y, dirX, dirY) {
+	didScroll (x, y, dirX, dirY) {
 		const
 			{dataSize} = this.props,
-			{firstIndex, numOfItems} = this.state,
+			{firstIndex} = this.state,
 			{isPrimaryDirectionVertical, threshold, dimensionToExtent, maxFirstIndex, scrollBounds} = this,
 			{gridSize} = this.primary,
 			maxPos = isPrimaryDirectionVertical ? scrollBounds.maxTop : scrollBounds.maxLeft,
@@ -620,22 +553,12 @@ class VirtualListCore extends Component {
 
 		if (firstIndex !== newFirstIndex) {
 			this.setState({firstIndex: newFirstIndex});
-		} else {
-			this.positionItems({updateFrom: firstIndex, updateTo: firstIndex + numOfItems});
-		}
-	}
-
-	applyStyleToExistingNode = (index, ...rest) => {
-		const node = this.getItemNode(index);
-
-		if (node) {
-			this.composeStyle(node.style, ...rest);
 		}
 	}
 
 	applyStyleToNewNode = (index, ...rest) => {
 		const
-			{component, data} = this.props,
+			{component, data, getNodeIndexToBeFocused, initItemRef} = this.props,
 			{numOfItems} = this.state,
 			key = index % numOfItems,
 			itemElement = component({
@@ -649,40 +572,11 @@ class VirtualListCore extends Component {
 		this.composeStyle(style, ...rest);
 
 		this.cc[key] = React.cloneElement(itemElement, {
-			ref: (index === this.nodeIndexToBeFocused) ? (ref) => this.initItemRef(ref, index) : null,
-			className: classNames(css.listItem, itemElement.props.className),
+			ref: (index === getNodeIndexToBeFocused()) ? (ref) => initItemRef(ref, index) : null,
+			className: classNames(cssItem.listItem, itemElement.props.className),
+			['data-preventscrollonfocus']: true, // Added this attribute to prevent scroll on focus by browser
 			style: {...itemElement.props.style, ...style}
 		});
-	}
-
-	focusOnNode = (node) => {
-		if (node) {
-			Spotlight.focus(node);
-		}
-	}
-
-	focusOnItem = (index) => {
-		const item = this.containerRef.querySelector(`[data-index='${index}'].spottable`);
-
-		if (Spotlight.isPaused()) {
-			Spotlight.resume();
-			this.forceUpdate();
-		}
-		this.focusOnNode(item);
-		this.nodeIndexToBeFocused = null;
-	}
-
-	initItemRef = (ref, index) => {
-		if (ref) {
-			this.focusOnItem(index);
-		}
-	}
-
-	focusByIndex = (index) => {
-		// We have to focus node async for now since list items are not yet ready when it reaches componentDid* lifecycle methods
-		setTimeout(() => {
-			this.focusOnItem(index);
-		}, 0);
 	}
 
 	applyStyleToHideNode = (index) => {
@@ -693,33 +587,35 @@ class VirtualListCore extends Component {
 		this.cc[key] = (<div {...attributes} />);
 	}
 
-	positionItems ({updateFrom, updateTo}) {
+	positionItems () {
 		const
 			{dataSize} = this.props,
-			{isPrimaryDirectionVertical, dimensionToExtent, primary, secondary, scrollPosition} = this;
-		let hideTo = 0;
+			{firstIndex, numOfItems} = this.state,
+			{isPrimaryDirectionVertical, dimensionToExtent, primary, secondary, cc} = this,
+			diff = firstIndex - this.lastFirstIndex,
+			updateFrom = (cc.length === 0 || 0 >= diff || diff >= numOfItems) ? firstIndex : this.lastFirstIndex + numOfItems;
+		let
+			hideTo = 0,
+			updateTo = (cc.length === 0 || -numOfItems >= diff || diff > 0) ? firstIndex + numOfItems : this.lastFirstIndex;
+
+		if (updateFrom >= updateTo) {
+			return;
+		} else if (updateTo > dataSize) {
+			hideTo = updateTo;
+			updateTo = dataSize;
+		}
 
 		// we only calculate position of the first child
 		let
 			{primaryPosition, secondaryPosition} = this.getGridPosition(updateFrom),
 			width, height;
 
-		primaryPosition -= scrollPosition;
 		width = (isPrimaryDirectionVertical ? secondary.itemSize : primary.itemSize) + 'px';
 		height = (isPrimaryDirectionVertical ? primary.itemSize : secondary.itemSize) + 'px';
 
-		if (updateTo > dataSize) {
-			hideTo = updateTo;
-			updateTo = dataSize;
-		}
-
 		// positioning items
 		for (let i = updateFrom, j = updateFrom % dimensionToExtent; i < updateTo; i++) {
-			if (this.updateFrom === null || this.updateTo === null || this.updateFrom > i || this.updateTo <= i) {
-				this.applyStyleToNewNode(i, width, height, primaryPosition, secondaryPosition);
-			} else {
-				this.applyStyleToExistingNode(i, width, height, primaryPosition, secondaryPosition);
-			}
+			this.applyStyleToNewNode(i, width, height, primaryPosition, secondaryPosition);
 
 			if (++j === dimensionToExtent) {
 				secondaryPosition = 0;
@@ -734,28 +630,44 @@ class VirtualListCore extends Component {
 			this.applyStyleToHideNode(i);
 		}
 
-		this.updateFrom = updateFrom;
-		this.updateTo = updateTo;
+		this.lastFirstIndex = firstIndex;
 	}
 
-	composeStyle (style, width, height, ...rest) {
+	scrollToPosition (x, y) {
+		const node = this.containerRef;
+
+		node.scrollTo((this.context.rtl && !this.isPrimaryDirectionVertical) ? this.scrollBounds.maxLeft - x : x, y);
+	}
+
+	composeStyle (style, width, height, primaryPosition, secondaryPosition) {
+		const {x, y} = this.getXY(primaryPosition, secondaryPosition);
+
 		if (this.isItemSized) {
 			style.width = width;
 			style.height = height;
 		}
 		style.position = 'absolute';
-		this.composeTransform(style, ...rest);
+
+		/* FIXME: RTL / this calculation only works for Chrome */
+		style.transform = 'translate(' + (this.context.rtl ? -x : x) + 'px,' + y + 'px)';
 	}
 
 	getXY = (primaryPosition, secondaryPosition) => {
-		const rtlDirection = this.context.rtl ? -1 : 1;
-		return (this.isPrimaryDirectionVertical ? {x: (secondaryPosition * rtlDirection), y: primaryPosition} : {x: (primaryPosition * rtlDirection), y: secondaryPosition});
+		return (this.isPrimaryDirectionVertical ? {x: secondaryPosition, y: primaryPosition} : {x: primaryPosition, y: secondaryPosition});
 	}
 
-	composeTransform (style, primaryPosition, secondaryPosition = 0) {
-		const {x, y} = this.getXY(primaryPosition, secondaryPosition);
+	updateMoreInfo (dataSize, primaryPosition) {
+		const
+			{dimensionToExtent, moreInfo} = this,
+			{itemSize, gridSize, clientSize} = this.primary;
 
-		style.transform = 'translate3d(' + x + 'px,' + y + 'px,0)';
+		if (dataSize <= 0) {
+			moreInfo.firstVisibleIndex = null;
+			moreInfo.lastVisibleIndex = null;
+		} else {
+			moreInfo.firstVisibleIndex = (Math.floor((primaryPosition - itemSize) / gridSize) + 1) * dimensionToExtent;
+			moreInfo.lastVisibleIndex = Math.min(dataSize - 1, Math.ceil((primaryPosition + clientSize) / gridSize) * dimensionToExtent - 1);
+		}
 	}
 
 	getScrollHeight = () => (this.isPrimaryDirectionVertical ? this.getVirtualScrollDimension() : this.scrollBounds.clientHeight)
@@ -776,7 +688,7 @@ class VirtualListCore extends Component {
 
 	calculatePositionOnFocus = ({item, scrollPosition = this.scrollPosition}) => {
 		const
-			{pageScroll} = this.props,
+			{lastFocusedIndex, pageScroll} = this.props,
 			{numOfItems} = this.state,
 			{primary} = this,
 			offsetToClientEnd = primary.clientSize - primary.itemSize,
@@ -785,17 +697,20 @@ class VirtualListCore extends Component {
 		if (!isNaN(focusedIndex)) {
 			let gridPosition = this.getGridPosition(focusedIndex);
 
-			if (numOfItems > 0 && focusedIndex % numOfItems !== this.lastFocusedIndex % numOfItems) {
-				this.getItemNode(this.lastFocusedIndex).blur();
+			if (numOfItems > 0 && focusedIndex % numOfItems !== lastFocusedIndex % numOfItems) {
+				const node = this.contentRef.children[this.lastFocusedIndex % numOfItems];
+				if (node) {
+					node.blur();
+				}
 			}
-			this.nodeIndexToBeFocused = null;
-			this.lastFocusedIndex = focusedIndex;
 
 			if (primary.clientSize >= primary.itemSize) {
 				if (gridPosition.primaryPosition > scrollPosition + offsetToClientEnd) { // forward over
 					gridPosition.primaryPosition -= pageScroll ? 0 : offsetToClientEnd;
 				} else if (gridPosition.primaryPosition >= scrollPosition) { // inside of client
-					gridPosition.primaryPosition = scrollPosition;
+					// This code uses the trick to change the target position slightly which will not affect the actual result
+					// since a browser ignore `scrollTo` method if the target position is same as the current position.
+					gridPosition.primaryPosition = scrollPosition + (this.scrollPosition === scrollPosition ? 0.1 : 0);
 				} else { // backward over
 					gridPosition.primaryPosition -= pageScroll ? offsetToClientEnd : 0;
 				}
@@ -805,301 +720,6 @@ class VirtualListCore extends Component {
 			// scrondaryPosition should be 0 here.
 			gridPosition.secondaryPosition = 0;
 			return this.gridPositionToItemPosition(gridPosition);
-		}
-	}
-
-	setRestrict = (bool) => {
-		Spotlight.set(this.props['data-container-id'], {restrict: (bool) ? 'self-only' : 'self-first'});
-	}
-
-	setSpotlightContainerRestrict = (keyCode, target) => {
-		const
-			{dataSize} = this.props,
-			{isPrimaryDirectionVertical, dimensionToExtent} = this,
-			index = Number.parseInt(target.getAttribute(dataIndexAttribute)),
-			canMoveBackward = index >= dimensionToExtent,
-			canMoveForward = index < (dataSize - (((dataSize - 1) % dimensionToExtent) + 1));
-		let isSelfOnly = false;
-
-		if (isPrimaryDirectionVertical) {
-			if (isUp(keyCode) && canMoveBackward || isDown(keyCode) && canMoveForward) {
-				isSelfOnly = true;
-			}
-		} else if (isLeft(keyCode) && canMoveBackward || isRight(keyCode) && canMoveForward) {
-			isSelfOnly = true;
-		}
-
-		this.setRestrict(isSelfOnly);
-	}
-
-	findSpottableItem = (indexFrom, indexTo) => {
-		const
-			{data, dataSize} = this.props,
-			safeIndexFrom = clamp(0, dataSize - 1, indexFrom),
-			safeIndexTo = clamp(-1, dataSize, indexTo),
-			delta = (indexFrom < indexTo) ? 1 : -1;
-
-		if (indexFrom < 0 && indexTo < 0 || indexFrom >= dataSize && indexTo >= dataSize) {
-			return -1;
-		}
-
-		if (safeIndexFrom !== safeIndexTo) {
-			for (let i = safeIndexFrom; i !== safeIndexTo; i += delta) {
-				if (data[i] && data[i].disabled === false) {
-					return i;
-				}
-			}
-		}
-
-		return -1;
-	}
-
-	getIndexToScrollDisabled = (direction, currentIndex) => {
-		const
-			{data, dataSize, spacing} = this.props,
-			{dimensionToExtent, primary, findSpottableItem} = this,
-			{firstVisibleIndex, lastVisibleIndex} = this.moreInfo,
-			numOfItemsInPage = (Math.floor((primary.clientSize + spacing) / primary.gridSize) * dimensionToExtent),
-			isPageDown = (direction === 'down' || direction === 'right') ? 1 : -1;
-		let candidateIndex = -1;
-
-		/* First, find a spottable item in this page */
-		if (isPageDown === 1) { // Page Down
-			if ((lastVisibleIndex - (lastVisibleIndex % dimensionToExtent || dimensionToExtent)) >= currentIndex) {
-				candidateIndex = findSpottableItem(
-					lastVisibleIndex,
-					currentIndex - (currentIndex % dimensionToExtent) + dimensionToExtent - 1
-				);
-			}
-		} else if (firstVisibleIndex + dimensionToExtent <= currentIndex) { // Page Up
-			candidateIndex = findSpottableItem(
-				firstVisibleIndex,
-				currentIndex - (currentIndex % dimensionToExtent)
-			);
-		}
-
-		/* Second, find a spottable item in the next page */
-		if (candidateIndex === -1) {
-			if (isPageDown === 1) { // Page Down
-				candidateIndex = findSpottableItem(lastVisibleIndex + numOfItemsInPage, lastVisibleIndex);
-			} else { // Page Up
-				candidateIndex = findSpottableItem(firstVisibleIndex - numOfItemsInPage, firstVisibleIndex);
-			}
-		}
-
-		/* Last, find a spottable item in a whole data */
-		if (candidateIndex === -1) {
-			if (isPageDown === 1) { // Page Down
-				candidateIndex = findSpottableItem(lastVisibleIndex + numOfItemsInPage + 1, dataSize);
-			} else { // Page Up
-				candidateIndex = findSpottableItem(firstVisibleIndex - numOfItemsInPage - 1, -1);
-			}
-		}
-
-		/* For grid lists, find the nearest item from the current item */
-		if (candidateIndex !== -1) {
-			const
-				currentPosInExtent = currentIndex % dimensionToExtent,
-				firstIndexInExtent = candidateIndex - (candidateIndex % dimensionToExtent),
-				lastIndexInExtent = clamp(firstIndexInExtent, dataSize - 1, firstIndexInExtent + dimensionToExtent);
-			let
-				minDistance = dimensionToExtent,
-				distance,
-				index;
-			for (let i = firstIndexInExtent; i <= lastIndexInExtent; ++i) {
-				if (data[i] && !data[i].disabled) {
-					distance = Math.abs(currentPosInExtent - i % dimensionToExtent);
-					if (distance < minDistance) {
-						minDistance = distance;
-						index = i;
-					}
-				}
-			}
-
-			return index;
-		} else {
-			return -1;
-		}
-	}
-
-	getIndexToScroll = (direction, currentIndex) => {
-		const
-			{dataSize, spacing} = this.props,
-			{dimensionToExtent, primary} = this,
-			numOfItemsInPage = Math.floor((primary.clientSize + spacing) / primary.gridSize) * dimensionToExtent,
-			factor = (direction === 'down' || direction === 'right') ? 1 : -1;
-		let indexToScroll = currentIndex + factor * numOfItemsInPage;
-
-		if (indexToScroll < 0) {
-			indexToScroll = currentIndex % dimensionToExtent;
-		} else if (indexToScroll >= dataSize) {
-			indexToScroll = dataSize - dataSize % dimensionToExtent + currentIndex % dimensionToExtent;
-			if (indexToScroll >= dataSize) {
-				indexToScroll = dataSize - 1;
-			}
-		}
-
-		return indexToScroll === currentIndex ? -1 : indexToScroll;
-	}
-
-	scrollToNextItem = ({direction, focusedItem}) => {
-		const
-			{data} = this.props,
-			focusedIndex = Number.parseInt(focusedItem.getAttribute(dataIndexAttribute)),
-			{firstVisibleIndex, lastVisibleIndex} = this.moreInfo;
-		let indexToScroll = -1;
-
-		if (Array.isArray(data) && data.some((item) => item.disabled)) {
-			indexToScroll = this.getIndexToScrollDisabled(direction, focusedIndex);
-		} else {
-			indexToScroll = this.getIndexToScroll(direction, focusedIndex);
-		}
-
-		if (indexToScroll !== -1) {
-			const
-				isRtl = this.context.rtl,
-				isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right');
-
-			// To prevent item positioning issue, make all items to be rendered.
-			this.updateFrom = null;
-			this.updateTo = null;
-
-			if (firstVisibleIndex <= indexToScroll && indexToScroll <= lastVisibleIndex) {
-				const node = this.containerRef.querySelector(`[data-index='${indexToScroll}'].spottable`);
-
-				if (node) {
-					Spotlight.focus(node);
-				}
-			} else {
-				// Scroll to the next spottable item without animation
-				if (!Spotlight.isPaused()) {
-					Spotlight.pause();
-				}
-				focusedItem.blur();
-			}
-			this.nodeIndexToBeFocused = this.lastFocusedIndex = indexToScroll;
-			this.props.cbScrollTo({index: indexToScroll, stickTo: isForward ? 'end' : 'start', animate: false});
-		}
-
-		return true;
-	}
-
-	shouldPreventScrollByFocus = () => this.isScrolledBy5way
-
-	jumpToSpottableItem = (keyCode, target) => {
-		const
-			{cbScrollTo, data, dataSize} = this.props,
-			{firstIndex, numOfItems} = this.state,
-			currentIndex = Number.parseInt(target.getAttribute(dataIndexAttribute));
-
-		if (!data || !Array.isArray(data) || !data[currentIndex] || data[currentIndex].disabled) {
-			return false;
-		}
-
-		const
-			isForward = (
-				this.isPrimaryDirectionVertical && isDown(keyCode) ||
-				!this.isPrimaryDirectionVertical && (!this.context.rtl && isRight(keyCode) || this.context.rtl && isLeft(keyCode)) ||
-				null
-			),
-			isBackward = (
-				this.isPrimaryDirectionVertical && isUp(keyCode) ||
-				!this.isPrimaryDirectionVertical && (!this.context.rtl && isLeft(keyCode) || this.context.rtl && isRight(keyCode)) ||
-				null
-			);
-
-		let nextIndex = -1;
-
-		if (isForward) {
-			// See if the next item is spottable then delegate scroll to onFocus handler
-			if (currentIndex < dataSize - 1 && !data[currentIndex + 1].disabled) {
-				return false;
-			}
-
-			for (let i = currentIndex + 2; i < dataSize; i++) {
-				if (!data[i].disabled) {
-					nextIndex = i;
-					break;
-				}
-			}
-
-			// If there is no item which could get focus forward,
-			// we need to set restriction option to `self-first`.
-			if (nextIndex === -1) {
-				this.setRestrict(false);
-			}
-		} else if (isBackward) {
-			// See if the next item is spottable then delegate scroll to onFocus handler
-			if (currentIndex > 0 && !data[currentIndex - 1].disabled) {
-				return false;
-			}
-
-			for (let i = currentIndex - 2; i >= 0; i--) {
-				if (!data[i].disabled) {
-					nextIndex = i;
-					break;
-				}
-			}
-
-			// If there is no item which could get focus backward,
-			// we need to set restriction option to `self-first`.
-			if (nextIndex === -1) {
-				this.setRestrict(false);
-			}
-		} else {
-			return false;
-		}
-
-		if (nextIndex !== -1 && (firstIndex > nextIndex || nextIndex >= firstIndex + numOfItems)) {
-			// When changing from "pointer" mode to "5way key" mode,
-			// a pointer is hidden and a last focused item get focused after 30ms.
-			// To make sure the item to be blurred after that, we used 50ms.
-			setTimeout(() => {
-				target.blur();
-			}, 50);
-
-			this.nodeIndexToBeFocused = this.lastFocusedIndex = nextIndex;
-
-			if (!Spotlight.isPaused()) {
-				Spotlight.pause();
-			}
-
-			cbScrollTo({
-				index: nextIndex,
-				stickTo: isForward ? 'end' : 'start'
-			});
-			return true;
-		}
-
-		return false;
-	}
-
-	onKeyDown = (e) => {
-		const {keyCode, target} = e;
-
-		this.isScrolledBy5way = false;
-		if (getDirection(keyCode)) {
-			this.setSpotlightContainerRestrict(keyCode, target);
-			this.isScrolledBy5way = this.jumpToSpottableItem(keyCode, target);
-		}
-		forwardKeyDown(e, this.props);
-	}
-
-	handleGlobalKeyDown = () => {
-		this.setContainerDisabled(false);
-	}
-
-	setContainerDisabled = (bool) => {
-		const containerNode = this.containerRef;
-
-		if (containerNode) {
-			containerNode.setAttribute(dataContainerDisabledAttribute, bool);
-
-			if (bool) {
-				document.addEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
-			} else {
-				document.removeEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
-			}
 		}
 	}
 
@@ -1127,13 +747,12 @@ class VirtualListCore extends Component {
 		if (clientWidth !== scrollBounds.clientWidth || clientHeight !== scrollBounds.clientHeight) {
 			this.calculateMetrics(props);
 			this.updateStatesAndBounds(props);
+			this.setContainerSize();
 			return true;
 		}
 
 		return false;
 	}
-
-	getItemNode = (index) => (this.itemContainerRef.children[index % this.state.numOfItems])
 
 	// render
 
@@ -1145,96 +764,59 @@ class VirtualListCore extends Component {
 
 	render () {
 		const
-			props = Object.assign({}, this.props),
-			{firstIndex, numOfItems} = this.state,
-			{primary, cc} = this;
+			{className, style, withPlaceholder, ...rest} = this.props,
+			{primary, cc} = this,
+			mergedClasses = classNames(css.list, this.containerClass, className);
 
-		delete props.cbScrollTo;
-		delete props.clientSize;
-		delete props.component;
-		delete props.data;
-		delete props.dataSize;
-		delete props.direction;
-		delete props.itemSize;
-		delete props.overhang;
-		delete props.pageScroll;
-		delete props.spacing;
+		delete rest.cbScrollTo;
+		delete rest.clientSize;
+		delete rest.component;
+		delete rest.data;
+		delete rest.dataSize;
+		delete rest.direction;
+		delete rest.getNodeIndexToBeFocused;
+		delete rest.itemSize;
+		delete rest.initItemRef;
+		delete rest.lastFocusedIndex;
+		delete rest.nodeIndexToBeFocused;
+		delete rest.overhang;
+		delete rest.pageScroll;
+		delete rest.preservedIndex;
+		delete rest.restoreLastFocused;
+		delete rest.spacing;
+		delete rest.withPlaceholder;
 
 		if (primary) {
-			this.positionItems({updateFrom: firstIndex, updateTo: firstIndex + numOfItems});
+			this.positionItems();
 		}
 
-		const needsScrollingPlaceholder = this.nodeIndexToBeFocused != null && Spotlight.isPaused();
-
 		return (
-			<div {...props} onKeyDown={this.onKeyDown} ref={this.initContainerRef}>
-				{cc.length ? <div ref={this.initItemContainerRef}>{cc}</div> : null}
-				{primary ? null : (
-					<SpotlightPlaceholder
-						data-index={0}
-						data-vl-placeholder
-						onFocus={this.handlePlaceholderFocus}
-						role="region"
-					/>
-				)}
-				{needsScrollingPlaceholder ? (
-					<SpotlightPlaceholder />
-				) : null}
+			<div className={mergedClasses} ref={this.initContainerRef} style={style}>
+				<div {...rest} onKeyDown={this.onKeyDown} ref={this.initContentRef}>
+					{withPlaceholder(cc.length ? cc : null)}
+				</div>
 			</div>
 		);
 	}
 }
 
 /**
- * {@link moonstone/VirtualList.VirtualListBase} is a base component for
+ * {@link moonstone/VirtualList.VirtualListBaseNative} is a base component for
  * {@link moonstone/VirtualList.VirtualList} and
  * {@link moonstone/VirtualList.VirtualGridList} with Scrollable and SpotlightContainerDecorator applied.
  *
- * @class VirtualListBase
+ * @class VirtualListBaseNative
  * @memberof moonstone/VirtualList
  * @mixes moonstone/Scrollable
  * @mixes spotlight/SpotlightContainerDecorator
  * @ui
  * @private
  */
-const VirtualListBase = SpotlightContainerDecorator(
-	{
-		enterTo: 'last-focused',
-		/*
-		 * Returns the data-index as the key for last focused
-		 */
-		lastFocusedPersist: (node) => {
-			const indexed = node.dataset.index ? node : node.closest('[data-index]');
-			if (indexed) {
-				return {
-					container: false,
-					element: true,
-					key: indexed.dataset.index
-				};
-			}
-		},
-		/*
-		 * Restores the data-index into the placeholder if its the only element. Tries to find a
-		 * matching child otherwise.
-		 */
-		lastFocusedRestore: ({key}, all) => {
-			if (all.length === 1 && 'vlPlaceholder' in all[0].dataset) {
-				all[0].dataset.index = key;
-
-				return all[0];
-			}
-
-			return all.reduce((focused, node) => {
-				return focused || node.dataset.index === key && node;
-			}, null);
-		},
-		preserveId: true,
-		restrict: 'self-first'
-	},
-	Scrollable(
-		VirtualListCore
+const VirtualListBaseNative = VirtualListContainerSpottable(
+	ScrollableNative( // including ScrollableSpotlightContainerDecorator
+		VirtualListCoreSpottable('Native')(VirtualListCoreNative)
 	)
 );
 
-export default VirtualListBase;
-export {gridListItemSizeShape, VirtualListCore, VirtualListBase};
+export default VirtualListBaseNative;
+export {gridListItemSizeShape, VirtualListCoreNative, VirtualListBaseNative};
